internal/buildscripts/packaging/release/requirements.txt
docker==4.3.1
PyGithub==1.40
requests==2.22.0